requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
python-dotenv
